+++ esm/index.js
@@ -5,3 +5,4 @@ export * from "./other";
 export * from "./param";
 export * from "./request";
 export * from "./time";
+export * from "./string";